old-tests/submission/W3C/domhtml/HTMLDocument10.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - DOCUMENT</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/dom.html#dom-document-forms" rel="help">
<link rel="author" title="NIST">
</head>
<body id="TEST-BODY"> 

<form action="/support/html/sample.html" method="post">
<p>
<textarea cols="20" name="text1" rows="7"></textarea>
<input name="submit1" type="submit" value="Submit">
<input name="submit2" type="reset" value="Reset">
</p>
</form>

<p>The forms attribute returns a collection of all the forms in a document.

    Retrieve the forms attribute from the document and examine its value.</p>


<script>
var domtest = async_test("The forms attribute returns a collection of all the forms in a document.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vforms;
      var vlength;
      var doc;
      doc = window.document;
      vforms = doc.forms;

      vlength = vforms.length;

      assert_equals(vlength,1);
       
    domtest.done();
}));

</script>
<div id="log"></div>
</body></html>
